#66, #67: Update readme and confirmation
Update the confirmation page to use the renamed verison of the OPC class
function that determines whether the order qualifies for free shipping
and modify the method used to determine if the order's total value has
changed.  The starting value is now passed via POST variable from the
checkout-one initial page.