Prevent "Free shipping over $0.00" when 0 threshold
When `free_free` has a minimum threshold, the displayed description outputs that amount. 
When the threshold is `$0.00` it shouldn't output the threshold.

Possible improvement, which requires other changes to language files, could include outputting a different description instead of suppressing it like this PR does.
One could consider pulling a define from another module, but that could get messy.