Avoid duplication of checkoutOneGuestInfo container
The restoreCustomerInfo endpoint returns a rendered opc_customer_info.php, which does not only include the inner contents of `div#checkoutOneGuestInfo`.
This is then inserted into the existing `div#checkoutOneGuestInfo`, causing an html structure similar to this:
```html
<div id="checkoutOneGuestInfo">
    <div id="checkoutOneGuestInfo">
```
This is obviously invalid, so this commit causes the existing one to be replaced instead.

There are still issues here around how the template also includes comments, and they're duplicated every time this (or what's there already) is run. This isn't great, but-
1. It's not a functional issue the way the previous one was for my CSS
2. It doesn't have an easy answer
so I don't care as much.